moved the files into folders
