+++ Categorize Password/Categorize Password.py
