+++ Jugglingalgorithm/Jugglingalgorithm.py
