+++ factorial/factorial.py
