+++ hacktober translator/hacktober translator.py
