+++ lcs/lcs.py
