+++ pattern19/pattern19.py
