+++ plaindrome/plaindrome.py
