+++ supportvectormachine/supportvectormachine.py
